app/order_management_microservice/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests==2.26.0